moved potentially flaky test into sequential test group
